estructura main y estructura de stack

+++ practica-git/src/stack.c
@@ -1,12 +1,18 @@
 #include "stack.h"
 
-
+// push a double to the stack
 void push(double data){
+	
 	return;
 }
-int pop(void){
+
+// pop a double to the stack
+double pop(void){
+	
 	return 0;
 }
+
+// check if empty
 bool empty(void){
 	return true;
 }
